Exchange opaque token with JWT
use the ldapauth traefik plugin as a token exchange.
Replaces an opaque token by a JWT.

Fixes a bug in the MockTCPServer which prevented clean closing of the
socket when the server got stopped.

Implements multi connection handling as well by running the message
handler in an own thread with each new accepted connection.

Tests the traefik plugin JWT retrieval with mocked memcache and ldap
server.