Make OAuth2 clients configurable
OAuth2 client_ids and corresponding redirect_uris need to be set in the
config file. To check whether the requested client id is registered,
the configured clients get searched for the provided client_id. Only,
if the client_id was found, the process continues. Searching for the
client_id in the list of objects required less code with the new slices
package. This package required an update of go.