Add simple equality verification provider
=== USE THIS PROVIDER ONLY FOR DEVELOPMENT! ===

Allows to configure a provider which is supported by the provider
selector to read data which will be checked for equality with passed in
data.

The verifyProvider interface can be used to define secret verificators
which secret source are defined by the supported data providers as it
uses internally the providerSelector.