Add Mapper to map IdP keys to claims
Allows to define mappers that get a key and value from the IdP passed
and return a claim and its value which will be added to the token.

The LDAP Entry Attributes get mapped to JWT Claims according to a
provided map. The LDAP Entry Attribute values stay as they are.

The mappers can be configured. As of now, only key-mappers are
supported. Script-based mappers to transform IdP values are imaginable.

Since mappers are client specific as the output values have to match the
client specific format, mappers get defined with the client
configuration.

Client configs should have a keys-mapper as first mapper to define the
keys with their values they require from the IdP.

The mappers interface allows to define different mapping implementations
that get called with different algorithms.

As of now, only a simple `SequentialMapper` algorithm is implemented
that takes a key/value pair and calls each defined mapping function in
order as they were listed in the clients configuration.

As of now, the OAuth2 clients mapper functions offer only a simple
key-mapper that replaces each input key with a configured replacement
key and keeps the value unmodified. The key-key-mapping can be defined
in the clients `mappers` config below the key `keys` like

```
oauth2:
  clients:
  - client_id: "client"
    [...]
    mappers:
    - keys:
        cn: "uid"
```

The example above maps the LDAP attribute `cn` of the authenticated user
to the key `uid` in the JWT when the client with the client_id `client`
gets requested.

In order to read the LDAP entry, it was required to use a search filter.
To mock the search filter in the tests (I know, better to use an own
interface, but mocking a LDAP server makes more fun :-)), the LDAP
protocol mocks need to get extended by supporting search filters. The
mock reads the LDAP requests (at least the required bytes) and responses
with a fake search result that got generated by ChatGPT.