Add EnvVarProvider
A provider that reads the content of a configured environment variable.
It can be used with the providerSelector. It returns an error if the
environment variable does not exist when the provider gets opened or
when the provider reads the environment variable. The read functions
reads always the current value of the environment variable.