Implement cache encryption of private claims
During the auth, the user gets authenticated with the IdP. At this
stage, the IdP has already the possibility to return the user entry. To
avoid accessing the IdP a second time during the token retrieval, the
user entry keys that the client requires get cached. To prevent
reading private claims as plaintext from the cache, the claims will get
AES encrypted if cache encryption is enabled.